+++ barc/src/lib.rs
@@ -73,21 +73,12 @@ pub const V2_MAX_HBLOCK: usize =        0xff_fff;
 /// 2<sup>40</sup> (1 TiB) - 1.
 pub const V2_MAX_REQ_BODY: u64 = 0xf_fff_fff_fff;
 
-/// Meta header name bytes for the complete URL used in the request.
-#[deprecated(since="0.5.0", note="use hname_meta_url() instead")]
-pub static META_URL: &[u8]             = b"url";
-
 /// Meta `HeaderName` for the complete URL used in the request.
 #[inline]
 pub fn hname_meta_url() -> http::header::HeaderName {
     static NAME: &str = "url";
     HeaderName::from_static(NAME)
 }
-
-/// Meta header name bytes for the HTTP method used in the request,
-/// e.g. "GET", "POST", etc.
-#[deprecated(since="0.5.0", note="use hname_meta_method() instead")]
-pub static META_METHOD: &[u8]          = b"method";
 
 /// Meta `HeaderName` for the HTTP method used in the request, e.g. "GET",
 /// "POST", etc.
@@ -97,11 +88,6 @@ pub fn hname_meta_method() -> http::header::HeaderName {
     HeaderName::from_static(NAME)
 }
 
-/// Meta header name bytes for the response version, e.g. "HTTP/1.1",
-/// "HTTP/2.0", etc.
-#[deprecated(since="0.5.0", note="use hname_meta_res_version() instead")]
-pub static META_RES_VERSION: &[u8]     = b"response-version";
-
 /// Meta `HeaderName` for the response version, e.g. "HTTP/1.1", "HTTP/2.0",
 /// etc.
 #[inline]
@@ -109,12 +95,6 @@ pub fn hname_meta_res_version() -> http::header::HeaderName {
     static NAME: &str = "response-version";
     HeaderName::from_static(NAME)
 }
-
-/// Meta header name bytes for the response numeric status code, SPACE,
-/// and then a standardized _reason phrase_, e.g. "200 OK". The later is
-/// intended only for human readers.
-#[deprecated(since="0.5.0", note="use hname_meta_res_status() instead")]
-pub static META_RES_STATUS: &[u8]      = b"response-status";
 
 /// Meta `HeaderName` for the response numeric status code, SPACE, and then a
 /// standardized _reason phrase_, e.g. "200 OK". The later is intended only
@@ -124,12 +104,6 @@ pub fn hname_meta_res_status() -> http::header::HeaderName {
     static NAME: &str = "response-status";
     HeaderName::from_static(NAME)
 }
-
-/// Meta header name bytes for a list of content or transfer encodings
-/// decoded for the current response body. The value is in HTTP
-/// content-encoding header format, e.g. "chunked, gzip".
-#[deprecated(since="0.5.0", note="use hname_meta_res_decoded() instead")]
-pub static META_RES_DECODED: &[u8]     = b"response-decoded";
 
 /// Meta `HeaderName` for a list of content or transfer encodings decoded for
 /// the current response body. The value is in HTTP content-encoding header